fix(sidebar): address code review feedback and apply orange theme
